+++ 5-logging/books/api.js
@@ -1,4 +1,4 @@
-// Copyright 2017, Google, Inc.
+// Copyright 2019, Google, Inc.
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
@@ -15,7 +15,7 @@
 
 const express = require('express');
 const bodyParser = require('body-parser');
-const model = require('./model-datastore');
+const model = require('./firestore');
 
 const router = express.Router();
