--- 5-logging/package.json
@@ -1,58 +1,0 @@
-{
-  "name": "nodejs-getting-started",
-  "description": "End to end sample for running Node.js applications on Google Cloud Platform",
-  "version": "1.0.0",
-  "license": "Apache-2.0",
-  "author": "Google Inc.",
-  "engines": {
-    "node": ">=8"
-  },
-  "repository": "https://github.com/GoogleCloudPlatform/nodejs-getting-started",
-  "main": "src/index.js",
-  "private": true,
-  "cloud-repo-tools": {
-    "requiresKeyFile": true,
-    "requiresProjectId": true,
-    "test": {
-      "app": {
-        "args": [
-          "app.js"
-        ],
-        "msg": "Bookshelf - Node.js on Google Cloud Platform"
-      }
-    }
-  },
-  "scripts": {
-    "start": "node app.js",
-    "test": "repo-tools test app && ava -t 30s --tap test/*.test.js",
-    "e2e": "repo-tools test deploy && E2E_TESTS=True ava -t 30s --tap test/*.test.js"
-  },
-  "dependencies": {
-    "@google-cloud/connect-datastore": "^3.0.0",
-    "@google-cloud/datastore": "^4.0.0",
-    "@google-cloud/debug-agent": "^4.0.0",
-    "@google-cloud/logging-winston": "^0.11.0",
-    "@google-cloud/storage": "^1.6.0",
-    "@google-cloud/trace-agent": "^4.0.0",
-    "body-parser": "^1.18.2",
-    "express": "^4.16.4",
-    "express-session": "^1.15.6",
-    "express-winston": "^2.5.0",
-    "lodash": "^4.17.11",
-    "multer": "^1.3.0",
-    "nconf": "^0.10.0",
-    "passport": "^0.4.0",
-    "passport-google-oauth20": "^2.0.0",
-    "prompt": "^1.0.0",
-    "pug": "^2.0.3",
-    "uglify-js": "^3.3.12",
-    "winston": "^2.4.0"
-  },
-  "devDependencies": {
-    "@google-cloud/nodejs-repo-tools": "^3.0.0",
-    "ava": "^0.25.0",
-    "proxyquire": "^2.0.0",
-    "supertest": "^4.0.0",
-    "sinon": "^7.0.0"
-  }
-}
